Import NIRS data

This function will import a .nir file, identify the baseline values and return a dataframe with the NIRS data in a long format.

import_nirs(nirFile)

Arguments

nirFile

Path to the .nir file

Value

A dataframe with the data in a long format with the following columns: t, startTime, fileName, dataRow, optode, freq, nirValue.

t: time in seconds from the start of the recording. You will see that this matches the values of the first column in COBI Studio .nir file.

startTime: the start time of the recording. This is taken from the .nir file.

fileName: the name of the file (full path)

dataRow: the row number of the data as it is in the original .nir file. The first row of the data is the first row after the baseline data ends in the original .nir file.

optode: the optode number (e.g., 1-16)

freq: the frequency (e.g., 730, 850)

nirValue: the value of the nir data (light intensity) for that optode and frequency at that time point. This matches the values in the original .nir file.

baselineValue: the baseline value for that optode and frequency. This is taken from the 10 second baseline period at the start of the recording, and is stored in the original .nir file. You can choose to use this value to calculate the change in optical density, if you wish. Either way, the baseline value is retained in the data frame for reference.

Details

This function will import a .nir file, identify the baseline values and return a dataframe with the data in a long format.

Examples

nirsData <- import_nirs("path/to/nir/file.nir")
#> Warning: cannot open file 'path/to/nir/file.nir': No such file or directory
#> Error in file(con, "r"): cannot open the connection